refactor: change `userId` type from `Long` to `Integer` in article and recruit controllers

+++ src/main/java/kr/ac/koreatech/sw/kosp/domain/community/article/model/Article.java
@@ -91,7 +91,7 @@ public class Article extends BaseEntity {
         return board.getId();
     }
 
-    public Long getAuthorId() {
+    public Integer getAuthorId() {
         return author.getId();
     }
